+++ src/main/java/org/simplestudio/controller/StepTplController.java
@@ -3,6 +3,7 @@ package org.simplestudio.controller;
 import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
+import org.simplestudio.dao.Step;
 import org.simplestudio.dao.StepTpl;
 import org.simplestudio.util.ConstantUtil;
 import org.simplestudio.util.StudioUtil;
@@ -52,21 +53,41 @@ public class StepTplController extends Controller {
 				"select *", expectSqlSelect);
 		renderJson(StudioUtil.convertPage2EasyUiPage(page));
 	}
-	
-	public void getStepTplList(){
+
+	public void getStepTplList() {
 		String sql = "select * from t_steptpl";
 		List<StepTpl> stepTplList = StepTpl.dao.find(sql);
 		renderJson(stepTplList);
 	}
 
-	public void delete(){
+	public void delete() {
 		String ids = getPara("ids");
-		if(StringUtils.isNotBlank(ids)){
+		String renderText = ConstantUtil.RESULT_SUCCESS;
+
+		if (StringUtils.isNotBlank(ids)) {
 			String[] idArr = ids.split(",");
-			for(int i=0;i<idArr.length;i++){
-				StepTpl.dao.deleteById(Long.parseLong(idArr[i]));
+			StringBuffer errorMsg = new StringBuffer();
+			for (int i = 0; i < idArr.length; i++) {
+				Step step = Step.dao.findFirst(
+						"select * from t_step where tpl_id=?", idArr[i]);
+				if (step != null) {
+					if (errorMsg.length() > 0) {
+						errorMsg.append(",");
+					}
+					StepTpl stepTpl = StepTpl.dao.findById(idArr[i]);
+					errorMsg.append("[").append(stepTpl.getStr("name"))
+							.append("]");
+				}
+			}
+			if (errorMsg.length() == 0) {
+				for (int i = 0; i < idArr.length; i++) {
+					StepTpl.dao.deleteById(Long.parseLong(idArr[i]));
+				}
+			} else {
+				renderText = "步骤模板" + errorMsg.toString()
+						+ "已被引用，无法删除！请先解除引用关系后再删除！";
 			}
 		}
-		renderText(ConstantUtil.RESULT_SUCCESS);
+		renderText(renderText);
 	}
 }
